pyproject.toml
[tool.poetry]
name = "atomic-habits-drf"
version = "0.1.0"
description = ""
authors = ["Denis_L <denis88lyapin@gmail.com>"]
readme = "README.md"
packages = [{include = "atomic_habits_drf"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.6"
djangorestframework = "^3.14.0"
psycopg2-binary = "^2.9.9"
django-filter = "^23.3"
djangorestframework-simplejwt = "^5.3.0"
drf-yasg = "^1.21.7"
celery = "^5.3.4"
redis = "^5.0.1"
django-celery-beat = "^2.5.0"
python-decouple = "^3.8"
ipython = "^8.16.1"
django-cors-headers = "^4.3.0"
requests = "^2.31.0"
coverage = "^7.3.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
